Simplify build: - don’t use templates anymore: generate valid rust file with constant - commit that constant file - compute at runtime some values
